# Image Optimization 성능 비교

커스텀 압축 구현과 `brower-image-compression` 압축의 성능을 테스트한 결과

## 성능 테스트 결과

|이미지 크기|압축 방식|처리 시간|압축 후|압축률|
|---|---|---|---|---|
|9.5MB|브라우저|1707.80ms|3033.06KB|68.27%|
||커스텀|1231.70ms|167.31KB|98.25%|
|277.52KB|브라우저|229.10ms|237.02KB|14.60%|
||커스텀|60.90ms|18.64KB|93.28%|
|165.22KB|브라우저|237.70ms|138.45KB|16.20%|
||커스텀|146.30ms|11.15KB|93.25%|
|474.71KB|브라우저|205.80ms|402.01KB|15.32%|
||커스텀|137.30ms|46.81KB|90.14%|

### ISSUE
![32312](https://github.com/user-attachments/assets/d7467078-c0c6-426d-b14e-1cefff81eeb0)

avif, webp 등 일부 형식의 이미지 업로드시 
network 400 error
sever -  code: 504 , message : '이미지  파일만 업로드 가능합니다.'

### 임시 대응 
모든 이미지 확장자를 JPEG로 변환 처리
- webp, avif 파일의 경우: JPEG 변환 시 기존 압축률 30% 감소
- 기타 이미지 파일: JPEG 변환 후 평균 압축률 50%