@bunzzeok 2024-12-13
![image](https://github.com/user-attachments/assets/25816671-14e0-46ba-8322-75de4218ef9d)

추가적으로 80번 issue와 동일한 문제 같음으로, 
close #80 

남겨주시면 issue close 가 될 거에요 !