@Minsoek96 2024-12-13
@bunzzeok 브릭스가 말씀하신 게 맞아요!!!
다만 현재 서버에서 webp, avif 등 일부 확장자를 차단하는 문제가 있어, 캡슐 생성 시 정상적으로
동작하지 않는 문제 떄문에 일시적으로 JPEG 변환을 적용해 두었습니다.
서버 문제가 해결되면 converToWebP만 WebP로 수정하면 될 것 같아요.

그리고 앞서 언급한 압축률은 webp 변환 테스트 결과였는데, 그 과정에서 모든 업로드 파일이 webp로 
변환하면서 위의 문제를 발견하게 되었어요. 제가 설명을 불명확하게 해서 오해의 소지가 있었네요. 미안합니다ㅎㅎ..

현재는 webp, avif파일의 경우 jpeg 변환하면 30% 정도 사이즈가 증가하지만 업로드는 정상적인걸로 확인했어요.
jpeg 에서 jpeg 같은 경우는 50% 정도 압축되는 것 같아요.!