@bunzzeok 2024-12-13
@Minsoek96 우선적으로 해당 부분은 백엔드가 완료될때까지 머지 할게요 !

80번 이슈는 완료해주시고, develop label 로 wepb 로 변환으로 이슈 하나 부탁드려요 ~ !